refactor(order): 의존성 방향 변경
