+++ src/main/java/kitchenpos/order/domain/OrderValidator.java
@@ -1,30 +1,6 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
-import org.springframework.stereotype.Component;
+public interface OrderValidator {
 
-import javax.persistence.EntityNotFoundException;
-
-@Component
-public class OrderValidator {
-
-    private final OrderTableRepository orderTableRepository;
-
-    public OrderValidator(OrderTableRepository orderTableRepository) {
-        this.orderTableRepository = orderTableRepository;
-    }
-
-    public void validate(Long orderTableId) {
-        validate(findOrderTableById(orderTableId));
-    }
-
-    void validate(OrderTable orderTable) {
-        orderTable.validateOrderTableChangable();
-    }
-
-    private OrderTable findOrderTableById(Long orderTableId) {
-        return orderTableRepository.findById(orderTableId)
-                .orElseThrow(EntityNotFoundException::new);
-    }
+    void validate(Long orderTableId);
 }
